requirements.txt
dnspython==2.6.2
requests==2.31.0
python-dotenv==1.0.0